Optimizations and Unit Testing for All Hash Set Implementations (#60)
* Feature: U64x2StaticHashSet
    - Refactored from U64x2HashSet to match the naming convention of other hash set implementations
    - Redesigned the declaration method to use a declarator macro matching other static hash set implementations
    - Create a helper method U64x2Equal in simd.h to replace U64x2HashSetInternalM128Equal
    - Replace U64x2HashSetInternalHash128to64 with the equivalent in core/data_structures/hash.h
    - Applied pointer hoisting optimizations
    - Added benchmarks

* Feature/Testing: TierPositionHashSet
    - Optimized using similar techniques as in other hash sets
    - Added benchmarks and unit tests
    - Updated Doxygen

* Feature: Implemented TierPositionStaticHashSet

* Feature: Implemented Int64StaticHashSet and wrapper type PositionStaticHashSet

* Chore: Replaced usages of PositionHashSet and TierPositionHashSet where the set size can be upper-bounded with PositionStaticHashSet and TierPositionStaticHashSet in Quixo

* Chore: Add missing symmetry enumeration optimization to Mills and Quixo

* Fix: Removed all timer pause and resume operations from benchmarks as they take much longer than the operations being benchmarked

* Fix: Integer overflow bug in simd_two_piece.h where the multiplication of two combinations may overflow 32-bits without type casting

+++ src/core/hash/simd_two_piece.h
@@ -349,8 +349,8 @@ void SimdTwoPieceHashContextDestroy(SimdTwoPieceHashContext *context);
 static inline int64_t SimdTwoPieceHashGetNumPositionsFixedTurn(
     const SimdTwoPieceHashContext *context, int num_x, int num_o) {
     const int32_t board_size = context->board_size;
-    return context->nCr[board_size - num_o][num_x] *
-           context->nCr[board_size][num_o];
+    return (int64_t)context->nCr[board_size - num_o][num_x] *
+           (int64_t)context->nCr[board_size][num_o];
 }
 
 /**
